add minor changes

+++ naf/WEB-INF/src/privilege/PrivilegeService.java
@@ -4,6 +4,8 @@ import java.util.List;
 
 import com.terapico.naf.spring.SpringBeanFactory;
 
+import test.MessageBox;
+
 public class PrivilegeService {
 	private UserDAOJDBCTemplateImpl impl;
 
@@ -31,11 +33,17 @@ public class PrivilegeService {
 		return impl.listUsers();
 	}
 
-	public List<User> 删除所有用户() {
-		impl.deleteAll();
+	public Object 删除所有用户(String 请输入WYSC) {
+		if(!请输入WYSC.equals("WYSC")){
+			
+			return new MessageBox("危险操作，输入不对，不予执行!");
+		}
+		
+		
+		int count=impl.deleteAll();
 		
 
-		return impl.listUsers();
+		return new MessageBox("成功删除"+count+"用户");
 	}
 	public List<User> 删除多个用户(Integer 编号之间以逗号分割[]) {
 		for (Integer id : 编号之间以逗号分割) {
